feat - #9 폴더링 구조 변경
