--- src/main/java/sopt/org/HMH/common/config/JpaAuditingConfig.java
@@ -1,9 +1,0 @@
-package sopt.org.HMH.common.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@Configuration
-@EnableJpaAuditing
-public class JpaAuditingConfig {
-}
