--- src/main/java/sopt/org/HMH/common/domain/BaseTimeEntity.java
@@ -1,22 +1,0 @@
-package sopt.org.HMH.common.domain;
-
-import jakarta.persistence.EntityListeners;
-import jakarta.persistence.MappedSuperclass;
-import java.time.LocalDateTime;
-import lombok.Getter;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-@MappedSuperclass
-@EntityListeners(AuditingEntityListener.class)
-@Getter
-public abstract class BaseTimeEntity {
-
-    @CreatedDate
-    protected LocalDateTime createdAt;
-
-    @LastModifiedDate
-    protected LocalDateTime updatedAt;
-
-}
