--- src/main/java/sopt/org/HMH/common/exception/GlobalError.java
@@ -1,30 +1,0 @@
-package sopt.org.HMH.common.exception;
-
-import lombok.AllArgsConstructor;
-import org.springframework.http.HttpStatus;
-import sopt.org.HMH.common.exception.base.ErrorBase;
-
-@AllArgsConstructor
-public enum GlobalError implements ErrorBase {
-
-    ;
-
-    private final HttpStatus status;
-    private final String errorMessage;
-
-    @Override
-    public int getHttpStatusCode() {
-        return status.value();
-    }
-
-    @Override
-    public HttpStatus getHttpStatus() {
-        return this.status;
-    }
-
-    @Override
-    public String getErrorMessage() {
-        return this.errorMessage;
-    }
-
-}
