--- src/main/java/sopt/org/HMH/common/exception/GlobalException.java
@@ -1,13 +1,0 @@
-package sopt.org.HMH.common.exception;
-
-import sopt.org.HMH.common.exception.base.ExceptionBase;
-
-/**
- * 전역적 혹은 시스템 전체 범위에서 발생하는 예외
- */
-public class GlobalException extends ExceptionBase {
-
-    public GlobalException(GlobalError error) {
-        super(error);
-    }
-}
