--- src/main/java/sopt/org/HMH/common/exception/GlobalSuccess.java
@@ -1,30 +1,0 @@
-package sopt.org.HMH.common.exception;
-
-import lombok.AllArgsConstructor;
-import org.springframework.http.HttpStatus;
-import sopt.org.HMH.common.exception.base.SuccessBase;
-
-@AllArgsConstructor
-public enum GlobalSuccess implements SuccessBase {
-
-    ;
-
-    private final HttpStatus status;
-    private final String successMessage;
-
-    @Override
-    public int getHttpStatusCode() {
-        return this.status.value();
-    }
-
-    @Override
-    public HttpStatus getHttpStatus() {
-        return this.status;
-    }
-
-    @Override
-    public String getSuccessMessage() {
-        return this.successMessage;
-    }
-
-}
