+++ src/main/java/sopt/org/HMH/domain/app/controller/AppController.java
@@ -1,0 +1,11 @@
+package sopt.org.HMH.domain.app.controller;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/api/v1/app")
+public class AppController {
+}
