+++ src/main/java/sopt/org/HMH/domain/app/dto/response/AppResponse.java
@@ -1,0 +1,8 @@
+package sopt.org.HMH.domain.app.dto.response;
+
+public record AppResponse(
+        String os,
+        String appCode,
+        Long goalTime
+) {
+}
