+++ src/main/java/sopt/org/HMH/domain/app/repository/AppRepository.java
@@ -1,0 +1,7 @@
+package sopt.org.HMH.domain.app.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import sopt.org.HMH.domain.app.domain.App;
+
+public interface AppRepository extends JpaRepository<App, Long> {
+}
