+++ src/main/java/sopt/org/HMH/domain/challenge/dto/request/ChallengeRequestDTO.java
@@ -1,0 +1,12 @@
+package sopt.org.HMH.domain.challenge.dto.request;
+
+import sopt.org.HMH.domain.app.dto.response.AppResponse;
+
+import java.util.List;
+
+public record ChallengeRequestDTO(
+        Integer period,
+        Long goalTime,
+        List<AppResponse> apps
+) {
+}
