+++ src/main/java/sopt/org/HMH/domain/challenge/dto/request/DayChallengeRequestDTO.java
@@ -1,0 +1,7 @@
+package sopt.org.HMH.domain.challenge.dto.request;
+
+// 안쓰일 것 같음
+public record DayChallengeRequestDTO(
+
+) {
+}
