+++ src/main/java/sopt/org/HMH/domain/challenge/dto/response/CreatedChallengeResponseDTO.java
@@ -1,0 +1,4 @@
+package sopt.org.HMH.domain.challenge.dto.response;
+
+public class CreatedChallengeResponseDTO {
+}
