+++ src/main/java/sopt/org/HMH/domain/user/User.java
@@ -1,0 +1,18 @@
+package sopt.org.HMH.domain.user;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+import sopt.org.HMH.global.common.domain.BaseTimeEntity;
+
+import static jakarta.persistence.GenerationType.IDENTITY;
+
+@Entity
+public class User extends BaseTimeEntity {
+
+    @Id
+    @GeneratedValue(strategy = IDENTITY)
+    @Column(name = "user_id")
+    private Long Id;
+}
