+++ src/main/java/sopt/org/HMH/global/common/Util.java
@@ -1,0 +1,14 @@
+package sopt.org.HMH.global.common;
+
+import java.security.Principal;
+
+import static java.util.Objects.isNull;
+
+public class Util {
+//    public static Long getUserId(Principal principal) {
+//        if (isNull(principal)) {
+//            throw new SecurityException(EMPTY_ACCESS_TOKEN.getMessage());
+//        }
+//        return Long.valueOf(principal.getName());
+//    }
+}
