+++ src/main/java/sopt/org/HMH/global/common/config/JpaAuditingConfig.java
@@ -1,0 +1,12 @@
+package sopt.org.HMH.global.common.config;
+
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfiguration;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@Configuration
+@EnableJpaAuditing
+@SpringBootApplication(exclude = SecurityAutoConfiguration.class)
+public class JpaAuditingConfig {
+}
