+++ src/main/java/sopt/org/HMH/global/common/exception/GlobalSuccess.java
@@ -1,0 +1,30 @@
+package sopt.org.HMH.global.common.exception;
+
+import lombok.AllArgsConstructor;
+import org.springframework.http.HttpStatus;
+import sopt.org.HMH.global.common.exception.base.SuccessBase;
+
+@AllArgsConstructor
+public enum GlobalSuccess implements SuccessBase {
+    SUCCESS_CREATE_CHALLENGE(HttpStatus.OK, "챌린지 생성 성공"),
+    ;
+
+    private final HttpStatus status;
+    private final String successMessage;
+
+    @Override
+    public int getHttpStatusCode() {
+        return this.status.value();
+    }
+
+    @Override
+    public HttpStatus getHttpStatus() {
+        return this.status;
+    }
+
+    @Override
+    public String getSuccessMessage() {
+        return this.successMessage;
+    }
+
+}
