+++ src/main/java/sopt/org/HMH/global/common/exception/base/ErrorBase.java
@@ -1,0 +1,11 @@
+package sopt.org.HMH.global.common.exception.base;
+
+import org.springframework.http.HttpStatus;
+
+public interface ErrorBase extends RootEnum {
+    int getHttpStatusCode();
+
+    HttpStatus getHttpStatus();
+
+    String getErrorMessage();
+}
